relationship has to be nullable
If a child and parent are linked via the API, the relationship is left as `null`. The only way to update that is through the front-end. Loading the data will then likely fail for many schools, given this is possible. (We are working on fix, but the database schema will still be possible to be null even following that fix.)